+++ editor/src/components/canvas/canvas-strategies/canvas-strategy-types.ts
@@ -10,9 +10,9 @@ import type { InsertionSubject } from '../../editor/editor-modes'
 import type { AllElementProps } from '../../editor/store/editor-state'
 import type { CanvasCommand } from '../commands/commands'
 import type { ActiveFrameAction } from '../commands/set-active-frames-command'
-import type { GridCellCoordinates } from '../controls/grid-controls'
 import type { StrategyApplicationStatus } from './interaction-state'
 import type { TargetGridCellData } from './strategies/grid-helpers'
+import type { GridCellCoordinates } from './strategies/grid-cell-bounds'
 
 // TODO: fill this in, maybe make it an ADT for different strategies
 export interface CustomStrategyState {
