+++ editor/src/components/editor/global-shortcuts.tsx
@@ -963,6 +963,7 @@ export function handleKeyDown(
           editor.allElementProps,
           editor.elementPathTree,
           editor.selectedViews,
+          { scale: editor.canvas.scale, offset: editor.canvas.realCanvasOffset },
         )
 
         if (commands.length === 0) {
